replaced option menu for syncing feed and account with pull to sync
