+++ simplified-app-shared/src/main/res/layout/catalog_feed_groups_list.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<android.support.v4.widget.SwipeRefreshLayout
   xmlns:android="http://schemas.android.com/apk/res/android"
+  android:id="@+id/swipe_refresh_layout"
   android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  android:background="@color/normal_background"
-  android:orientation="vertical">
+  android:layout_height="match_parent">
 
   <ListView
     android:id="@+id/catalog_feed_blocks_list"
@@ -12,4 +11,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
 
-</LinearLayout>
+</android.support.v4.widget.SwipeRefreshLayout>
