+++ simplified-app-shared/src/main/res/layout/catalog_feed_nogroups.xml
@@ -14,19 +14,27 @@
     android:layout_width="match_parent"
     android:layout_height="1dp" />
 
-  <GridView
-    android:id="@+id/catalog_feed_nogroups_grid"
+  <android.support.v4.widget.SwipeRefreshLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/swipe_refresh_layout"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginStart="8dp"
-    android:layout_marginEnd="8dp"
-    android:paddingTop="8dp"
-    android:paddingBottom="8dp"
-    android:clipToPadding="false"
-    android:scrollbars="none"
-    android:verticalSpacing="8dp"
-    android:horizontalSpacing="8dp"
-    android:numColumns="@integer/catalog_feed_nogroups_columns">
-  </GridView>
+    android:layout_height="match_parent">
+
+    <GridView
+      android:id="@+id/catalog_feed_nogroups_grid"
+      android:layout_width="match_parent"
+      android:layout_height="match_parent"
+      android:layout_marginStart="8dp"
+      android:layout_marginEnd="8dp"
+      android:paddingTop="8dp"
+      android:paddingBottom="8dp"
+      android:clipToPadding="false"
+      android:scrollbars="none"
+      android:verticalSpacing="8dp"
+      android:horizontalSpacing="8dp"
+      android:numColumns="@integer/catalog_feed_nogroups_columns">
+    </GridView>
+
+  </android.support.v4.widget.SwipeRefreshLayout>
 
 </LinearLayout>
